refactor(frontend/copilot): address CodeRabbit nitpick comments
- Remove dead code branches in getAccordionMeta for both CreateAgent
  and EditAgent (isAgentSavedOutput and error fallthrough are
  unreachable since ToolAccordion is already guarded)
- Remove unused WarningDiamondIcon imports from both files
- Use index-based key instead of action.label in ToolErrorCard
- Move ClarifyingQuestion interface and normalizeClarifyingQuestions
  to helpers.tsx for better separation of concerns
