+++ autogpt_platform/frontend/src/app/(platform)/copilot/tools/CreateAgent/components/ClarificationQuestionsCard.tsx
@@ -7,44 +7,7 @@ import { Text } from "@/components/atoms/Text/Text";
 import { cn } from "@/lib/utils";
 import { ChatTeardropDotsIcon, CheckCircleIcon } from "@phosphor-icons/react";
 import { useEffect, useRef, useState } from "react";
-
-export interface ClarifyingQuestion {
-  question: string;
-  keyword: string;
-  example?: string;
-}
-
-export function normalizeClarifyingQuestions(
-  questions: Array<{ question: string; keyword: string; example?: unknown }>,
-): ClarifyingQuestion[] {
-  const seen = new Set<string>();
-
-  return questions.map((q, index) => {
-    let keyword = q.keyword?.trim().toLowerCase() || "";
-    if (!keyword) {
-      keyword = `question-${index}`;
-    }
-
-    let unique = keyword;
-    let suffix = 1;
-    while (seen.has(unique)) {
-      unique = `${keyword}-${suffix}`;
-      suffix++;
-    }
-    seen.add(unique);
-
-    const item: ClarifyingQuestion = {
-      question: q.question,
-      keyword: unique,
-    };
-    const example =
-      typeof q.example === "string" && q.example.trim()
-        ? q.example.trim()
-        : null;
-    if (example) item.example = example;
-    return item;
-  });
-}
+import type { ClarifyingQuestion } from "../helpers";
 
 interface Props {
   questions: ClarifyingQuestion[];
