+++ autogpt_platform/frontend/src/app/(platform)/copilot/tools/CreateAgent/helpers.tsx
@@ -157,3 +157,41 @@ export function truncateText(text: string, maxChars: number): string {
   if (trimmed.length <= maxChars) return trimmed;
   return `${trimmed.slice(0, maxChars).trimEnd()}…`;
 }
+
+export interface ClarifyingQuestion {
+  question: string;
+  keyword: string;
+  example?: string;
+}
+
+export function normalizeClarifyingQuestions(
+  questions: Array<{ question: string; keyword: string; example?: unknown }>,
+): ClarifyingQuestion[] {
+  const seen = new Set<string>();
+
+  return questions.map((q, index) => {
+    let keyword = q.keyword?.trim().toLowerCase() || "";
+    if (!keyword) {
+      keyword = `question-${index}`;
+    }
+
+    let unique = keyword;
+    let suffix = 1;
+    while (seen.has(unique)) {
+      unique = `${keyword}-${suffix}`;
+      suffix++;
+    }
+    seen.add(unique);
+
+    const item: ClarifyingQuestion = {
+      question: q.question,
+      keyword: unique,
+    };
+    const example =
+      typeof q.example === "string" && q.example.trim()
+        ? q.example.trim()
+        : null;
+    if (example) item.example = example;
+    return item;
+  });
+}
